LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/request-headers.any-expected.txt


PASS Fetch with GET
FAIL Fetch with HEAD assert_equals: Request should have header content-length: null expected (object) null but got (string) "0"
PASS Fetch with PUT without body
PASS Fetch with PUT with body
PASS Fetch with POST without body
PASS Fetch with POST with text body
PASS Fetch with POST with FormData body
PASS Fetch with POST with URLSearchParams body
FAIL Fetch with POST with Blob body assert_equals: Request should have header content-type: null expected (object) null but got (string) "application/x-www-form-urlencoded"
FAIL Fetch with POST with ArrayBuffer body assert_equals: Request should have header content-type: null expected (object) null but got (string) "application/x-www-form-urlencoded"
FAIL Fetch with POST with Uint8Array body assert_equals: Request should have header content-type: null expected (object) null but got (string) "application/x-www-form-urlencoded"
FAIL Fetch with POST with Int8Array body assert_equals: Request should have header content-type: null expected (object) null but got (string) "application/x-www-form-urlencoded"
FAIL Fetch with POST with Float32Array body assert_equals: Request should have header content-type: null expected (object) null but got (string) "application/x-www-form-urlencoded"
FAIL Fetch with POST with Float64Array body assert_equals: Request should have header content-type: null expected (object) null but got (string) "application/x-www-form-urlencoded"
FAIL Fetch with POST with DataView body assert_equals: Request should have header content-type: null expected (object) null but got (string) "application/x-www-form-urlencoded"
PASS Fetch with POST with Blob body with mime type
FAIL Fetch with Chicken assert_equals: Request should have header content-length: null expected (object) null but got (string) "0"
PASS Fetch with Chicken with body
PASS Fetch with GET and mode "cors" does not need an Origin header
PASS Fetch with POST and mode "same-origin" needs an Origin header
PASS Fetch with POST and mode "no-cors" needs an Origin header
PASS Fetch with PUT and mode "same-origin" needs an Origin header
PASS Fetch with TacO and mode "same-origin" needs an Origin header
PASS Fetch with TacO and mode "cors" needs an Origin header
